requirements.txt
pyyaml
bip44
python-dotenv
web3
eth-account
black